Refactored code. #145
