+++ src/Classes/IoHelper.php
@@ -1,20 +1,21 @@
 <?php
 
-namespace SleekDB\Traits;
+namespace SleekDB\Classes;
 
 use Closure;
+use Exception;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use SleekDB\Exceptions\IOException;
 use SleekDB\Exceptions\JsonException;
 
-trait IoHelperTrait {
+class IoHelper {
 
   /**
    * @param string $path
    * @throws IOException
    */
-  private static function _checkWrite(string $path)
+  public static function _checkWrite(string $path)
   {
     if(file_exists($path) === false){
       $path = dirname($path);
@@ -31,7 +32,7 @@ trait IoHelperTrait {
    * @param string $path
    * @throws IOException
    */
-  private static function _checkRead(string $path)
+  public static function _checkRead(string $path)
   {
     // Check if PHP has read permission
     if (!is_readable($path)) {
@@ -46,7 +47,8 @@ trait IoHelperTrait {
    * @return string
    * @throws IOException
    */
-  private static function getFileContent(string $filePath){
+  public static function getFileContent(string $filePath): string
+  {
 
     self::_checkRead($filePath);
 
@@ -69,7 +71,12 @@ trait IoHelperTrait {
     return $content;
   }
 
-  private static function writeContentToFile(string $filePath, string $content){
+  /**
+   * @param string $filePath
+   * @param string $content
+   * @throws IOException
+   */
+  public static function writeContentToFile(string $filePath, string $content){
 
     self::_checkWrite($filePath);
 
@@ -84,7 +91,8 @@ trait IoHelperTrait {
    * @return bool
    * @throws IOException
    */
-  private static function deleteFolder(string $folderPath){
+  public static function deleteFolder(string $folderPath): bool
+  {
     self::_checkWrite($folderPath);
     $it = new RecursiveDirectoryIterator($folderPath, RecursiveDirectoryIterator::SKIP_DOTS);
     $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
@@ -103,7 +111,7 @@ trait IoHelperTrait {
    * @param string $folderPath
    * @throws IOException
    */
-  private static function createFolder(string $folderPath){
+  public static function createFolder(string $folderPath){
     self::_checkWrite($folderPath);
     // Check if the data_directory exists or create one.
     if (!file_exists($folderPath) && !mkdir($folderPath, 0777, true) && !is_dir($folderPath)) {
@@ -115,12 +123,12 @@ trait IoHelperTrait {
 
   /**
    * @param string $filePath
-   * @param Closure $updateContentFunction
+   * @param Closure $updateContentFunction Has to return a string or an array that will be encoded to json.
    * @return string
    * @throws IOException
    * @throws JsonException
    */
-  private static function updateFileContent(string $filePath, Closure $updateContentFunction): string
+  public static function updateFileContent(string $filePath, Closure $updateContentFunction): string
   {
     self::_checkRead($filePath);
     self::_checkWrite($filePath);
@@ -162,25 +170,27 @@ trait IoHelperTrait {
    * @param string $filePath
    * @return bool
    */
-  private static function deleteFile(string $filePath){
+  public static function deleteFile(string $filePath): bool
+  {
 
     if(false === file_exists($filePath)){
       return true;
     }
     try{
       self::_checkWrite($filePath);
-    }catch(\Exception $exception){
+    }catch(Exception $exception){
       return false;
     }
 
-    return @unlink($filePath) && !file_exists($filePath);
+    return (@unlink($filePath) && !file_exists($filePath));
   }
 
   /**
    * @param array $filePaths
    * @return bool
    */
-  private static function deleteFiles(array $filePaths){
+  public static function deleteFiles(array $filePaths): bool
+  {
     foreach ($filePaths as $filePath){
       // if a file does not exist, we do not need to delete it.
       if(true === file_exists($filePath)){
@@ -189,11 +199,21 @@ trait IoHelperTrait {
           if(false === @unlink($filePath) || file_exists($filePath)){
             return false;
           }
-        } catch (\Exception $exception){
+        } catch (Exception $exception){
           return false;
         }
       }
     }
     return true;
+  }
+
+  /**
+   * Strip string for secure file access.
+   * @param string $string
+   * @return string
+   */
+  public static function secureStringForFileAccess(string $string): string
+  {
+    return (str_replace(array(".", "/", "\\"), "", $string));
   }
 }
