+++ src/Classes/NestedHelper.php
@@ -1,0 +1,110 @@
+<?php
+
+
+namespace SleekDB\Classes;
+
+
+use SleekDB\Exceptions\InvalidArgumentException;
+
+class NestedHelper
+{
+
+  /**
+   * Get nested properties of a store object.
+   * @param string $fieldName
+   * @param array $data
+   * @return mixed
+   * @throws InvalidArgumentException
+   */
+  public static function getNestedValue(string $fieldName, array $data)
+  {
+    $fieldName = trim($fieldName);
+    if (empty($fieldName)) {
+      throw new InvalidArgumentException('fieldName is not allowed to be empty');
+    }
+    // Dive deep step by step.
+    foreach (explode('.', $fieldName) as $i) {
+      // If the field does not exists we return null;
+      if (!isset($data[$i])) {
+        return null;
+      }
+      // The index is valid, collect the data.
+      $data = $data[$i];
+    }
+    return $data;
+  }
+
+  public static function updateNestedValue(string $fieldName, array &$data, $newValue){
+    $fieldNameArray = explode(".", $fieldName);
+    $value = $newValue;
+    if(count($fieldNameArray) > 1){
+      $data = self::_updateNestedValue($fieldNameArray, $data, $newValue, count($fieldNameArray));
+      return;
+    }
+    $data[$fieldNameArray[0]] = $value;
+  }
+
+  /**
+   * @param array $keysArray
+   * @param $data
+   * @param $newValue
+   * @param int $originalKeySize
+   * @return mixed
+   */
+  private static function _updateNestedValue(array $keysArray, $data, $newValue, int $originalKeySize)
+  {
+    if(empty($keysArray)){
+      return $newValue;
+    }
+    $currentKey = $keysArray[0];
+    $result = (is_array($data)) ? $data : [];
+    if(!is_array($data) || !array_key_exists($currentKey, $data)){
+      $result[$currentKey] = self::_updateNestedValue(array_slice($keysArray, 1), $data, $newValue, $originalKeySize);
+      if(count($keysArray) !== $originalKeySize){
+        return $result;
+      }
+    }
+    $result[$currentKey] = self::_updateNestedValue(array_slice($keysArray, 1), $data[$currentKey], $newValue, $originalKeySize);
+    return $result;
+  }
+
+  public static function createNestedArray(string $fieldName, $fieldValue): array
+  {
+    $temp = [];
+    $fieldNameArray = explode('.', $fieldName);
+    $fieldNameArrayReverse = array_reverse($fieldNameArray);
+    foreach ($fieldNameArrayReverse as $index => $i) {
+      if($index === 0){
+        $temp = array($i => $fieldValue);
+      } else {
+        $temp = array($i => $temp);
+      }
+    }
+
+    return $temp;
+  }
+
+  public static function removeNestedField(array &$document, string $fieldToRemove){
+    if (array_key_exists($fieldToRemove, $document)) {
+      unset($document[$fieldToRemove]);
+      return;
+    }
+    // should be a nested array at this point
+    $temp = &$document;
+    $fieldNameArray = explode('.', $fieldToRemove);
+    $fieldNameArrayCount = count($fieldNameArray);
+    foreach ($fieldNameArray as $index => $i) {
+      // last iteration
+      if(($fieldNameArrayCount - 1) === $index){
+        if(is_array($temp) && array_key_exists($i, $temp)) {
+          unset($temp[$i]);
+        }
+        break;
+      }
+      if(!is_array($temp) || !array_key_exists($i, $temp)){
+        break;
+      }
+      $temp = &$temp[$i];
+    }
+  }
+}
