+++ src/Exceptions/InvalidPropertyAccessException.php
@@ -2,6 +2,10 @@
 
 namespace SleekDB\Exceptions;
 
+/**
+ * Class InvalidPropertyAccessException
+ * @deprecated since version 2.7.
+ */
 class InvalidPropertyAccessException extends \Exception
 {
 }
